+++ packages/dive_ui/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Next
+
+- We have added Example 15, which demonstrates how to stream both video and audio while simultaneously recording using multiple cameras. The example also includes a mix of camera feeds and shows how to switch between them.
+- Fixed the use of obs-ffmpeg-mux for recording.
+
 ## 0.1.0
 
 - Initial release.